requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2.7
django-stubs-ext==4.2.5
django-types==0.19.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
executing==2.0.1
idna==3.4
importlib-resources==6.1.1
inflection==0.5.1
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.1
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.13.0
types-psycopg2==2.9.21.16
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.11
zipp==3.17.0
